Add unreliable car test
